requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
ChatterBot==1.0.1
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.4.4
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
MarkupSafe==1.1.1
mathparse==0.1.2
nltk==3.7
oauthlib==3.1.0
packaging==21.3
Pint==0.18
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.3
pyparsing==3.0.8
python-dateutil==2.7.5
pytz==2022.1
PyYAML==3.13
regex==2022.3.15
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
SQLAlchemy==1.2.19
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.8.0
